add DDIM with api cleanups and basic documentation
