+++ configs/ddim/cifar10.yaml
@@ -4,7 +4,7 @@ trainer:
   logger:
     class_path: pytorch_lightning.loggers.WandbLogger
     init_args:
-      project: "CIFAR10 Image Generation"
+      project: "CIFAR10_Image_Generation"
       name: "DDIM"
   enable_checkpointing: true
   callbacks:
@@ -70,26 +70,24 @@ trainer:
   inference_mode: true
 ckpt_path: null
 model:
-  class_path: LitDDIM
+  class_path: dmme.ddim.LitDDIM
   init_args:
     model:
       class_path: dmme.ddpm.UNet
       init_args:
         in_channels: 3
-        dim: 128
         pos_dim: 128
         emb_dim: 512
-        multipliers:
-          - 1
-          - 2
-          - 2
-          - 2
+        channels:
+          - 128
+          - 256
+          - 256
+          - 256
         attn_depth:
           - 2
         groups: 32
-        dropout: 0.1
+        drop_rate: 0.1
     tau_schedule: "linear"
-    eta: 1.0
     lr: 2e-4
     warmup: 5000
     imgsize:
@@ -100,7 +98,7 @@ model:
     timesteps: 1000
     decay: 0.9999
 data:
-  class_path: CIFAR10
+  class_path: dmme.data.CIFAR10
   init_args:
     data_dir: "."
     batch_size: 128
