+++ docs/source/data.md
@@ -1,0 +1,19 @@
+# Data
+
+```{eval-rst}
+.. currentmodule:: dmme.data
+
+.. autosummary::
+    :nosignatures:
+
+    DataModule
+    CIFAR10
+```
+
+```{toctree}
+:caption: 'Data:'
+:maxdepth: 2
+
+data/datamodule.md
+data/cifar10
+```
