+++ docs/source/data/cifar10.md
@@ -1,0 +1,8 @@
+# CIFAR10
+
+```{eval-rst}
+.. currentmodule:: dmme.data
+
+.. autoclass:: CIFAR10
+    :members:
+```
