+++ docs/source/data/datamodule.md
@@ -1,0 +1,10 @@
+# DataModule
+
+```{eval-rst}
+.. currentmodule:: dmme.data
+
+.. autoclass:: DataModule
+    :members:
+
+.. autofunction:: cpu_count
+```
