--- docs/source/datamodules.md
@@ -1,17 +1,0 @@
-# DataModules
-
-```{eval-rst}
-.. currentmodule:: dmme.data_modules
-
-.. autosummary::
-    :nosignatures:
-
-    Dataset
-    CIFAR10
-```
-```{toctree}
-:caption: 'DataModules:'
-:maxdepth: 2
-
-datamodules/cifar10
-```
