--- docs/source/datamodules/cifar10.md
@@ -1,11 +1,0 @@
-# CIFAR10
-
-```{eval-rst}
-.. currentmodule:: dmme
-
-.. autoclass:: CIFAR10
-    :members:
-
-.. currentmodule:: dmme.data_modules.dataset
-.. autofunction:: cpu_count
-```
