--- scripts/trainer.py
@@ -1,11 +1,0 @@
-from pytorch_lightning.cli import LightningCLI
-
-import dmme
-
-
-def main():
-    cli = LightningCLI()
-
-
-if __name__ == "__main__":
-    main()
