+++ src/dmme/data/__init__.py
@@ -1,0 +1,2 @@
+from .cifar10 import CIFAR10
+from .data_module import DataModule, cpu_count
