--- src/dmme/data_modules/__init__.py
@@ -1,5 +1,0 @@
-from .dataset import Dataset
-
-from .cifar10 import CIFAR10
-
-__all__ = ["Dataset", "CIFAR10"]
