--- src/dmme/ddim/__init__.py
@@ -1,4 +1,0 @@
-from .ddim import DDIMReverseProcess, DDIMSampler
-from .lit_ddim import LitDDIM
-
-__all__ = ["DDIMReverseProcess", "DDIMSampler"]
