+++ src/dmme/lr_scheduler/__init__.py
@@ -1,3 +1,1 @@
 from .warmup import WarmupLR
-
-__all__ = ["WarmupLR"]
