+++ src/dmme/trainer.py
@@ -1,7 +1,5 @@
 from pytorch_lightning.cli import LightningCLI
 
-import dmme
-
 
 def main():
     cli = LightningCLI(seed_everything_default=1337)
